[PRMDR-167] Change location of test data file to match other branch, put MockResponse class as a separate helper file
