+++ lambdas/tests/unit/helpers/mock_response.py
@@ -1,0 +1,11 @@
+class MockResponse:
+    def __init__(self, status_code, json_data):
+        self.status_code = status_code
+        self.json_data = json_data
+
+    def json(self):
+        return self.json_data
+
+    def content(self):
+        return repr(self.json_data)
+
