+++ lambdas/tests/unit/services/test_ods_api_service.py
@@ -2,7 +2,8 @@ import pytest
 
 from enums.permitted_role import PermittedRole
 from services.ods_api_service import OdsApiService
-from tests.unit.test_data.utils import load_ods_response_data
+from tests.unit.helpers.data.ods.utils import load_ods_response_data
+from tests.unit.helpers.mock_response import MockResponse
 from utils.exceptions import OrganisationNotFoundException, OdsErrorException
 
 
@@ -15,14 +16,18 @@ def test_fetch_organisation_data_valid_returns_organisation_data():
     assert "PrimaryRoleId" in str(actual)
 
 
-def test_fetch_organisation_data_404_raise_OrganisationNotFoundException():
-    test_ods_code = "non-exist-ods-code"
+def test_fetch_organisation_data_404_raise_OrganisationNotFoundException(mocker):
+    response_404 = MockResponse(404, {"errorCode": 404, "errorText": "Not found"})
 
+    mocker.patch("requests.get", return_value=response_404)
     with pytest.raises(OrganisationNotFoundException):
-        OdsApiService.fetch_organisation_data(test_ods_code)
+        OdsApiService.fetch_organisation_data("non-exist-ods-code")
 
 
-def test_fetch_organisation_data_catch_all_raises_OdsErrorException():
+def test_fetch_organisation_data_catch_all_raises_OdsErrorException(mocker):
+    response_400 = MockResponse(400, "BadRequest")
+
+    mocker.patch("requests.get", return_value=response_400)
     invalid_ods_code = "!@?£?$?@?"
 
     with pytest.raises(OdsErrorException):
@@ -31,6 +36,7 @@ def test_fetch_organisation_data_catch_all_raises_OdsErrorException():
 
 @pytest.fixture()
 def mock_ods_responses():
+    # load test data from several json files and pass to below tests in a dict
     yield load_ods_response_data()
